# docs/swagger.yaml
openapi: 3.1.0
info:
  title: 1337Park
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/watchguard-475/1337Park/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:3000/api/
  description: localhost
...
paths:
...
  /users/{userid}:
    delete:
      tags:
      - user
      summary: delete profile
      operationId: deleteProfile
      parameters:
      - name: userid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
    patch:
      tags:
      - user
      summary: Update Profile
      description: customer can update his profile ONLY but admin can update certain fields of any user
      operationId: updateProfile
      parameters:
      - name: userid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
...
  /users:
    get:
      tags:
      - user
      summary: Get all users
      description: admin can get all users
      operationId: getAllUsers
      parameters:
      - name: userId
        in: query
        description: Optional userid
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
...
components:
  schemas:
...
    UserDTO:
      title: UserDTO
      type: object
      properties:
        id:
          type: string
          x-stoplight:
            id: k1fjywcjyn2xh
        name:
          type: string
        email:
          type: string
        officeName:
          type: string
        role:
          type: string
          x-stoplight:
            id: bph5ntil01gtp
          enum:
          - Customer
          - Admin
      description: ""
...
    Error:
      title: Error
      type: object
      properties:
        Message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT